% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summary.R
\name{project_summary}
\alias{project_summary}
\title{Create a project summary HTML page}
\usage{
project_summary(..., number_sections = FALSE)
}
\arguments{
\item{...}{Options passed through to \code{\link[bookdown]{html_document2}}}
}
\value{
R Markdown output format
}
\description{
Consistent starting point for Data School participants creating
a project summary page.
}
\examples{
render("input.Rmd", project_summary())
}
